latest

+++ gokul/week1/lect-1.4/1.4.10.py
@@ -5,5 +5,5 @@
 # ```
 
 a = input(int())
-b = int(a * a)
+b = a*a
 print(b)
